requirements.txt
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
vk-api==11.9.9

selenium~=4.4.3